requirements.txt
PyStemmer
psutil==5.9.6